+++ extension/src/Settings/Settings.ts
@@ -63,6 +63,10 @@ export class Settings {
   public enableTroubleshootingCommands = true;
   public useDictionaryInDTSImport = true;
   public enableXliffCache = true;
+  public preferLockedTranslations = true;
+  public translationFilenamePattern = "*.xlf";
+  // Other extension's settings:
+  public packageCachePath = undefined;
 
   constructor(workspaceFolderPath: string) {
     this.workspaceFolderPath = workspaceFolderPath;
@@ -130,6 +134,7 @@ export class LaunchSettings {
 
 export interface IAppSourceCopSettings {
   mandatoryAffixes: string[];
+  mandatoryPrefix: string;
 }
 
 export interface IExtensionPackage {