+++ extension/src/Settings/SettingsMap.ts
@@ -63,4 +63,9 @@ export const settingsMap = new Map<string, keyof Settings>([
   ["NAB.EnableTelemetry", "enableTelemetry"],
   ["NAB.EnableTroubleshootingCommands", "enableTroubleshootingCommands"],
   ["NAB.EnableXliffCache", "enableXliffCache"],
+  ["NAB.PreferLockedTranslations", "preferLockedTranslations"],
+  ["NAB.TranslationFilenamePattern", "translationFilenamePattern"],
+
+  // Other extension's settings:
+  ["al.packageCachePath", "packageCachePath"],
 ]);