+++ extension/src/test/Template.test.ts
@@ -4,6 +4,7 @@ import * as assert from "assert";
 import { TemplateSettings } from "../Template/TemplateTypes";
 import * as TemplateFunctions from "../Template/TemplateFunctions";
 import * as FileFunctions from "../FileFunctions";
+import _ = require("lodash");
 
 const templateSettingsFilename = "al.template.json";
 const sourceResourcesPath = "../../src/test/resources/templateSettings";
@@ -207,10 +208,7 @@ suite("Template", function () {
     assert.ok(appName, "AppName not ok");
     assert.strictEqual(
       path.basename(workspaceFile),
-      `${FileFunctions.replaceIllegalFilenameCharacters(
-        appName.replace(/ /g, "-"),
-        "-"
-      )}.code-workspace`,
+      "AppNameWithIllegalFilenameCharacters.code-workspace",
       "Unexpected workspace file name"
     );
 
@@ -220,7 +218,9 @@ suite("Template", function () {
           encoding: "utf8",
         })
         .replace(/[\r\n]*/g, ""),
-      `${appName}${templateSettings.mappings[1].value}${appName}${templateSettings.mappings[1].value}${templateSettings.mappings[2].value}`,
+      `${appName}${templateSettings.mappings[1].value}${appName}${
+        templateSettings.mappings[1].value
+      }${templateSettings.mappings[2].value}${_.kebabCase(appName)}`,
       "Unexpected content in file1"
     );
     assert.strictEqual(